refactor: 경로 변경 & 설명 추가 (#6)
