+++ src/main/java/teamficial/teamficial_be/domain/auth/controller/AuthController.java
@@ -24,6 +24,7 @@ public class AuthController {
     }
 
     @PostMapping("/auth/google")
+    @Operation(summary = "구글 로그인",description = "인가 코드를 통해 토큰을 발급받는 카카오 로그인 API입니다.")
     public ApiResponse<LoginResponseDTO.LoginTokenResponseDto> googleLogin(@RequestParam("code")String accessCode, @RequestParam("redirectUri") String redirectUri){
         return ApiResponse.onSuccess(authService.googleLogin(accessCode,redirectUri));
     }
