+++ src/main/java/teamficial/teamficial_be/global/security/google/GoogleUtil.java
@@ -10,7 +10,6 @@ import org.springframework.stereotype.Component;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestClient;
-import teamficial.teamficial_be.global.security.google.dto.GoogleDTO;
 
 import java.net.URLDecoder;
 import java.nio.charset.StandardCharsets;
